+++ src/libraries/SwapHelpers.sol
@@ -7,12 +7,7 @@ import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {IUniswapV2Router02} from "../interfaces/IUniswapV2Router02.sol";
 
 library SwapHelpers {
-    
-    function encodePath(address[] memory tokens, uint24[] memory fees)
-        public
-        pure
-        returns (bytes memory path)
-    {
+    function encodePath(address[] memory tokens, uint24[] memory fees) public pure returns (bytes memory path) {
         require(tokens.length == fees.length + 1, "Invalid input arrays");
 
         for (uint256 i = 0; i < fees.length; i++) {
@@ -23,7 +18,6 @@ library SwapHelpers {
         // Append the final token address
         path = abi.encodePacked(path, tokens[tokens.length - 1]);
     }
-
 
     function swapVersion3(
         ISwapRouter uniswapRouter,
@@ -38,7 +32,7 @@ library SwapHelpers {
             path: encodePath(path, fees),
             recipient: recipient,
             deadline: block.timestamp + 300,
-            amountIn:  amountIn,
+            amountIn: amountIn,
             amountOutMinimum: amountOutMinimum
         });
 
@@ -51,8 +45,8 @@ library SwapHelpers {
         uint256 amountIn,
         uint256 amountOutMin,
         address recipient
-    ) internal returns (uint amountOut) {
-        uint[] memory v2AmountOut = uniswapRouter.getAmountsOut(amountIn, path);
+    ) internal returns (uint256 amountOut) {
+        uint256[] memory v2AmountOut = uniswapRouter.getAmountsOut(amountIn, path);
         IERC20(path[0]).approve(address(uniswapRouter), amountIn);
         uniswapRouter.swapExactTokensForTokens(
             amountIn, //amountIn
@@ -80,5 +74,4 @@ library SwapHelpers {
             amountOut = swapTokensV2(uniswapRouterV2, path, amountIn, amountOutMinimum, recipient);
         }
     }
-
 }
