+++ src/orderManager/OrderManager.sol
@@ -63,10 +63,7 @@ contract OrderManager is Initializable, OwnableUpgradeable {
         _;
     }
 
-    function initialize(
-        address _usdcAddress,
-        address _indexFactory
-    ) external initializer {
+    function initialize(address _usdcAddress, address _indexFactory) external initializer {
         __Ownable_init(msg.sender);
         usdcAddress = _usdcAddress;
         factory = IndexFactory(_indexFactory);
@@ -98,21 +95,11 @@ contract OrderManager is Initializable, OwnableUpgradeable {
         orderNonceInfo.orderNonce += 1;
     }
 
-    function _transferInputTokenFromCaller(
-        address _inputToken,
-        uint256 _amount
-    ) internal {
-        require(
-            _inputToken != address(0),
-            "OrderManger: invalid token address"
-        );
+    function _transferInputTokenFromCaller(address _inputToken, uint256 _amount) internal {
+        require(_inputToken != address(0), "OrderManger: invalid token address");
         require(_amount > 0, "OrderManger: amount must be greater than 0");
 
-        IERC20(_inputToken).safeTransferFrom(
-            msg.sender,
-            address(this),
-            _amount
-        );
+        IERC20(_inputToken).safeTransferFrom(msg.sender, address(this), _amount);
     }
 
     function _initializeOrder(CreateOrderConfig memory _config) internal {
@@ -143,16 +130,11 @@ contract OrderManager is Initializable, OwnableUpgradeable {
         }
     }
 
-    function createOrder(
-        CreateOrderConfig memory _config
-    ) external onlyOperator returns (uint256 orderNonce) {
+    function createOrder(CreateOrderConfig memory _config) external onlyOperator returns (uint256 orderNonce) {
         // increasing order nonce
         _increaseOrderNonce(_config.isBuyOrder);
         // transfer USDC from caller to order manager contract
-        _transferInputTokenFromCaller(
-            _config.inputTokenAddress,
-            _config.inputTokenAmount
-        );
+        _transferInputTokenFromCaller(_config.inputTokenAddress, _config.inputTokenAmount);
         // initialize the order based on buy or sell
         _initializeOrder(_config);
         //call the provider function
@@ -172,10 +154,7 @@ contract OrderManager is Initializable, OwnableUpgradeable {
     }
 
     function completeOrder(uint256 _orderNonce) external onlyOperator {
-        require(
-            _orderNonce > 0 && _orderNonce <= orderNonceInfo.orderNonce,
-            "OrderManager: invalid order nonce"
-        );
+        require(_orderNonce > 0 && _orderNonce <= orderNonceInfo.orderNonce, "OrderManager: invalid order nonce");
         OrderInfo storage order = orderInfo[_orderNonce];
         require(!order.isExecuted, "OrderManager: order already executed");
         order.isExecuted = true;
@@ -186,15 +165,11 @@ contract OrderManager is Initializable, OwnableUpgradeable {
         uint256 _issuanceNonce,
         address _indexToken,
         address _underlyingTokenAddress,
-        uint _oldTokenValue,
-        uint _newTokenValue
+        uint256 _oldTokenValue,
+        uint256 _newTokenValue
     ) external onlyOperator {
         factory.handleCompleteIssuance(
-            _issuanceNonce,
-            _indexToken,
-            _underlyingTokenAddress,
-            _oldTokenValue,
-            _newTokenValue
+            _issuanceNonce, _indexToken, _underlyingTokenAddress, _oldTokenValue, _newTokenValue
         );
     }
 
@@ -204,11 +179,6 @@ contract OrderManager is Initializable, OwnableUpgradeable {
         address _underlyingTokenAddress,
         uint256 _outputValue
     ) external onlyOperator {
-        factory.handleCompleteRedemption(
-            _redemptionNonce,
-            _indexToken,
-            _underlyingTokenAddress,
-            _outputValue
-        );
+        factory.handleCompleteRedemption(_redemptionNonce, _indexToken, _underlyingTokenAddress, _outputValue);
     }
 }
